requirements.txt
streamlit>=1.32
torch>=2.0
torchvision>=0.15
Pillow>=9.0
jupyter